Create rough Data Dig Astronauts
My first time creating a case on my own from a set of already made data. I found it quite a challenge to come up with the queries and the basis of a case. 
I finally started to get it. I am understanding SQL more and more and although I thought maybe I am not smart enough. Im just going to keep pushing through.
on to lesson 1.4